Move strategies into a sandbox folder
