+++ src/ext/java/org/opentripplanner/ext/vehicletostopheuristics/BikeToStopSkipEdgeStrategy.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.algorithm.astar.strategies;
+package org.opentripplanner.ext.vehicletostopheuristics;
 
 import java.util.Collection;
 import java.util.Set;
@@ -6,6 +6,7 @@ import java.util.function.Function;
 import org.opentripplanner.model.BikeAccess;
 import org.opentripplanner.model.Stop;
 import org.opentripplanner.model.Trip;
+import org.opentripplanner.routing.algorithm.astar.strategies.SkipEdgeStrategy;
 import org.opentripplanner.routing.api.request.RoutingRequest;
 import org.opentripplanner.routing.core.State;
 import org.opentripplanner.routing.graph.Edge;
